backend/requirements.txt
Django==5.0.4
python-dotenv==1.1.0
python_jose==3.4.0
djangorestframework==3.14.0
psycopg2==2.9.9
django-cors-headers==4.3.1
Requests==2.32.3